Move the `altrep-rep-utils.c` functions to `altrep-rep.c`
