+++ src/altrep-rep.c
@@ -32,6 +32,19 @@ SEXP vctrs_is_vctrs_compact_rep(SEXP x) {
   return Rf_ScalarLogical(vec_is_vctrs_compact_rep(x));
 }
 
+
+// [[ register() ]]
+SEXP vctrs_is_vctrs_compact_rep_compact(SEXP x) {
+  if (!vec_is_vctrs_compact_rep(x)) {
+    Rf_errorcall(R_NilValue, "Internal error: `x` is not an ALTREP compact rep");
+  }
+
+  bool out = VCTRS_COMPACT_REP_IS_COMPACT(x);
+
+  return Rf_ScalarLogical(out);
+}
+
+
 // [[ init() ]]
 void vctrs_init_vctrs_compact_rep(DllInfo* dll) {
   vctrs_init_vctrs_compact_rep_int(dll);
@@ -54,6 +67,14 @@ SEXP vctrs_is_vctrs_compact_rep(SEXP x) {
   Rf_errorcall(R_NilValue, "Need R 3.5+ for ALTREP support");
   return R_NilValue;
 }
+
+
+// [[ register() ]]
+SEXP vctrs_is_vctrs_compact_rep_compact(SEXP x) {
+  Rf_errorcall(R_NilValue, "Need R 3.5+ for ALTREP support");
+  return R_NilValue;
+}
+
 
 // For DllInfo on R < 3.4
 #include <R_ext/Rdynload.h>
